add ChordController
