+++ src/AppBundle/Controller/SongController.php
@@ -7,9 +7,29 @@ use AppBundle\Form\Type\SongType;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template as Template;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 class SongController extends Controller
 {
+    /**
+     * Render front page
+     *
+     * @return Response
+     */
+    public function indexAction()
+    {
+        $songs = $this->get('doctrine_mongodb.odm.document_manager')
+            ->getRepository('AppBundle:Song')
+            ->findAll();
+
+        return $this->render(
+            'AppBundle::index.html.twig',
+            [
+                'songs' => $songs
+            ]
+        );
+    }
+
     /**
      * Method that addnew song
      *
