+++ src/AppBundle/Resources/config/routing.yml
@@ -1,23 +1,23 @@
 app:
     pattern:  /
-    defaults: { _controller: AppBundle:Default:index }
-
-app_videos_of_album:
-    pattern: /album/{id}
-    defaults: { _controller: AppBundle:Default:songsOfCategory}
+    defaults: { _controller: AppBundle:Song:index }
 
 app_video_addVideo:
     pattern: /new
     defaults: { _controller: AppBundle:Video:addVideo}
 
-#app_video_copyVideoToUser:
-#    pattern: /copyVideo/{user_id}/{video_id}
-#    defaults: { _controller: AppBundle:Default:copyVideoToUser}
+app_video_copyVideoToUser:
+    pattern: /copyVideo/{user_id}/{video_id}
+    defaults: { _controller: AppBundle:Video:copyVideoToUser}
 
 app_video_show:
     pattern: /showVideo
     defaults: { _controller: AppBundle:Video:showVideo}
 
+app_chord_show:
+    pattern: /showChord
+    defaults: { _controller: AppBundle:Chord:index}
+
 app_video_addSong:
     pattern: /addSong
-    defaults: { _controller: AppBundle:Default:addSong}
+    defaults: { _controller: AppBundle:Song:addSong}
