+++ src/AppBundle/Resources/views/menu.html.twig
@@ -21,6 +21,11 @@
         </li>
 
         <li>
+            <a class="blog-nav-item" href="{{ path('app_chord_show') }}">
+                {% trans %}ShowChord{% endtrans %}</a>
+        </li>
+
+        <li>
             <a class="blog-nav-item" href="{{ path('app_video_addSong') }}">
                 {% trans %}AddSong{% endtrans %}</a>
         </li>
